refactor(autoware_gyro_odometer): move the diagnostics decision into its own module
separated diagnostics function from logic because it's not
ROS-dependent but it's considered to be different from pure logic.
