+++ localization/autoware_gyro_odometer/CMakeLists.txt
@@ -5,6 +5,7 @@ find_package(autoware_cmake REQUIRED)
 autoware_package()
 
 ament_auto_add_library(${PROJECT_NAME} SHARED
+  src/gyro_odometer_diagnostics.cpp
   src/gyro_odometer_fusion.cpp
 )
