+++ localization/autoware_gyro_odometer/src/gyro_odometer_core.cpp
@@ -14,6 +14,7 @@
 
 #include "gyro_odometer_core.hpp"
 
+#include "gyro_odometer_diagnostics.hpp"
 #include "gyro_odometer_fusion.hpp"
 
 #include <rclcpp/rclcpp.hpp>
