+++ localization/autoware_gyro_odometer/src/gyro_odometer_fusion.cpp
@@ -17,12 +17,9 @@
 #include <autoware_utils_geometry/msg/covariance.hpp>
 #include <rclcpp/time.hpp>
 
-#include <fmt/core.h>
-
 #include <algorithm>
 #include <cmath>
 #include <deque>
-#include <string>
 
 namespace autoware::gyro_odometer
 {
@@ -116,46 +113,6 @@ geometry_msgs::msg::TwistWithCovarianceStamped apply_stop_compensation(
     result.twist.twist.angular.x = 0.0;
     result.twist.twist.angular.y = 0.0;
     result.twist.twist.angular.z = 0.0;
-  }
-
-  return result;
-}
-
-DiagnosticsResult determine_diagnostics(const DiagnosticsState & state)
-{
-  using diagnostic_msgs::msg::DiagnosticStatus;
-
-  DiagnosticsResult result;
-
-  const auto raise = [&result](const int8_t level, const std::string & message) {
-    result.entries.push_back({level, message});
-    result.level = std::max(result.level, level);
-    result.log_message += message;
-    result.log_message += "; ";
-  };
-
-  if (!state.vehicle_twist_arrived) {
-    raise(DiagnosticStatus::WARN, "Twist msg has not been arrived yet.");
-  }
-  if (!state.imu_arrived) {
-    raise(DiagnosticStatus::WARN, "IMU msg has not been arrived yet.");
-  }
-  if (state.latest_vehicle_twist_dt > state.message_timeout_sec) {
-    const std::string message = fmt::format(
-      "Vehicle twist msg is timeout. vehicle_twist_dt: {}[sec], tolerance {}[sec]",
-      state.latest_vehicle_twist_dt, state.message_timeout_sec);
-    raise(DiagnosticStatus::ERROR, message);
-  }
-  if (state.latest_imu_dt > state.message_timeout_sec) {
-    const std::string message = fmt::format(
-      "IMU msg is timeout. imu_dt: {}[sec], tolerance {}[sec]", state.latest_imu_dt,
-      state.message_timeout_sec);
-    raise(DiagnosticStatus::ERROR, message);
-  }
-  if (!state.is_succeed_transform_imu) {
-    raise(
-      DiagnosticStatus::ERROR,
-      "Please publish TF from " + state.output_frame + " to frame of IMU.");
   }
 
   return result;
