+++ localization/autoware_gyro_odometer/src/gyro_odometer_fusion.hpp
@@ -15,15 +15,11 @@
 #ifndef GYRO_ODOMETER_FUSION_HPP_
 #define GYRO_ODOMETER_FUSION_HPP_
 
-#include <diagnostic_msgs/msg/diagnostic_status.hpp>
 #include <geometry_msgs/msg/twist_with_covariance_stamped.hpp>
 #include <sensor_msgs/msg/imu.hpp>
 
 #include <array>
-#include <cstdint>
 #include <deque>
-#include <string>
-#include <vector>
 
 namespace autoware::gyro_odometer
 {
@@ -54,44 +50,6 @@ geometry_msgs::msg::TwistWithCovarianceStamped fuse_twist(
 /// otherwise the input is returned unchanged. Pure function returning a new message.
 geometry_msgs::msg::TwistWithCovarianceStamped apply_stop_compensation(
   const geometry_msgs::msg::TwistWithCovarianceStamped & twist_with_cov);
-
-/// \brief A single diagnostics finding: a severity level paired with its human-readable message.
-struct DiagnosticsEntry
-{
-  int8_t level{diagnostic_msgs::msg::DiagnosticStatus::OK};
-  std::string message;
-};
-
-/// \brief Inputs needed to decide the gyro_odometer diagnostics level and messages.
-struct DiagnosticsState
-{
-  bool vehicle_twist_arrived{false};
-  bool imu_arrived{false};
-  bool is_succeed_transform_imu{false};
-  double latest_vehicle_twist_dt{0.0};
-  double latest_imu_dt{0.0};
-  double message_timeout_sec{0.0};
-  std::string output_frame;
-};
-
-/// \brief Result of evaluating the diagnostics state.
-///
-/// \c entries holds one element per triggered condition (level + message), preserving the node's
-/// historical check order. \c level is the maximum severity across all entries. \c log_message is
-/// every entry message concatenated, each terminated by "; ".
-struct DiagnosticsResult
-{
-  std::vector<DiagnosticsEntry> entries;
-  int8_t level{diagnostic_msgs::msg::DiagnosticStatus::OK};
-  std::string log_message;
-};
-
-/// \brief Evaluate the diagnostics state into per-condition entries, an aggregated level, and a
-/// concatenated log message.
-///
-/// Pure function: no node, clock, or interface dependency. The messages are byte-for-byte identical
-/// to the node's original strings (the TF-failure message embeds \c state.output_frame).
-DiagnosticsResult determine_diagnostics(const DiagnosticsState & state);
 
 }  // namespace autoware::gyro_odometer
