+++ localization/autoware_gyro_odometer/test/test_gyro_odometer_fusion.cpp
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include "gyro_odometer_diagnostics.hpp"
 #include "gyro_odometer_fusion.hpp"
 
 #include <autoware_utils_geometry/msg/covariance.hpp>
